refactor(redaction): extract shared sensitive-key masking into _redaction
Add _redaction exporting SENSITIVE_KEYS, normalize_key and mask_value, and
import it into both RedactionFilter and ColorFormatter instead of each
module keeping its own key set and inline masking. ColorFormatter now masks
against the single unified key set (broader, safer) using the same
normalization rules. _filters re-exports _MASK/_DEFAULT_SENSITIVE_KEYS/
_normalize_key for backward compatibility.

Part of #207
